refactor payload config export
