--- apps/nextjs/src/payload.config.ts
@@ -1,1 +1,0 @@
-export { config as default } from "@acme/payload";
