+++ packages/api/package.json
@@ -30,6 +30,7 @@
     "@trpc/client": "^11.0.0-rc.466",
     "@trpc/react-query": "^11.0.0-rc.466",
     "@trpc/server": "^11.0.0-rc.466",
+    "payload": "beta",
     "superjson": "2.2.1",
     "zod": "catalog:"
   },
@@ -40,7 +41,6 @@
     "@types/node": "catalog:",
     "@types/react": "catalog:react18",
     "eslint": "catalog:",
-    "payload": "beta",
     "prettier": "catalog:",
     "typescript": "catalog:"
   },
