+++ packages/payload/package.json
@@ -5,8 +5,7 @@
   "type": "module",
   "exports": {
     ".": {
-      "types": "./dist/index.d.ts",
-      "default": "./src/index.ts"
+      "default": "./src/payload.config.ts"
     },
     "./env": "./src/env.ts"
   },
