+++ packages/payload/src/seed/index.ts
@@ -1,9 +1,9 @@
-import { getPayload } from "../index";
-
+import { getPayload } from "payload";
+import config from "@payload-config";
 import { seedUsers } from "./collections/users";
 
 async function run() {
-  const payload = await getPayload();
+  const payload = await getPayload({config});
 
   try {
     const isSeeded =
