+++ packages/payload/tsconfig.json
@@ -1,5 +1,10 @@
 {
   "extends": "@acme/tsconfig/internal-package.json",
+  "compilerOptions": {
+    "paths": {
+      "@payload-config": ["./src/payload.config.ts"]
+    }
+  },
   "include": ["src"],
   "exclude": ["node_modules"]
 }
